Dev

See merge request martind/nominate-your-bee!140

+++ nominate_app/views/nomination_index.py
@@ -16,81 +16,87 @@ class NominationIndexView(View):
 	def get(self, request):
 		comment_form = CommentForm()
 		sort = 'latest'
+		date_filter = None
 		page = request.GET.get('page', 1)
+
 		if 'Awards' in request.GET:
-			award_id = request.GET['Awards']
-			template_id = request.GET['Templates']
+			award_name = request.GET['Awards']
+			template_name = request.GET['Templates']
+			from_ = request.GET['from_']
+			to = request.GET['to']
 			sort = request.GET['Sort']
-			nominations = nomination_filter(award_id, template_id)
-			nominate_form = NominationFilterForm(initial={'Awards': award_id, 'Templates': template_id, 'Sort': sort})
+			if from_ and to:
+				date_filter = {
+					'from': datetime.strptime(from_, '%m/%d/%Y').date(),
+					'to': datetime.strptime(to, '%m/%d/%Y').date(),
+				}
+
+			nominate_form = NominationFilterForm(initial={'Awards': award_name, 'Templates': template_name, 'Sort': sort, 'from_': from_, 'to': to})
+			submissions =  submission_filter(award_name, template_name, date_filter, sort)
 
 		else:
 			nominate_form = NominationFilterForm()
-			nominations = Nomination.objects.all()
+			submissions = NominationSubmitted.objects.all()
+
+		paginator = Paginator(submissions, 10)
+		submissions = paginator.page(page)
+
+		instances = []
+		for submission in submissions:
+			instances.append(nomination_instance_detail(submission, request.user))
 
 
-		instance_details = []
-		instances = nomination_sort(nominations, sort)
-
-		for instance in instances:
-			instance_details.append({
-				'instance': instance,
-				'detail': nomination_instance_detail(instance, request.user)
-				})
-
-		paginator = Paginator(instance_details, 10)
-		instance_details = paginator.page(page)
-
-		return render(request, self.template_name, {'instances': instance_details, 'comment_form': comment_form, 'filter_form': nominate_form})
+		return render(request, self.template_name, {'instances': instances, 'comment_form': comment_form, 'filter_form': nominate_form})
 
 
 
-def nomination_filter(award_id, template_id):
-	awards = Awards.objects.all() if award_id == 'ALL' else Awards.objects.filter(id=int(award_id))
-	award_templates = AwardTemplate.objects.filter(award__in=awards) if template_id =='ALL' else AwardTemplate.objects.filter(award__in=awards, id=int(template_id))
-	nominations = Nomination.objects.filter(award_template__in=award_templates)
+def submission_filter(award_name, template_name, date_filter, sort):
+	submissions = NominationSubmitted.objects.all()
+	if award_name != 'ALL':
+		submissions = submissions.filter(award_name=award_name)
 
-	return nominations
+	if template_name != 'ALL':
+		submissions = submissions.filter(template_name=template_name)
 
+	if date_filter:
+		submissions = submissions.filter(submitted_at__gte=date_filter['from'], submitted_at__lte=date_filter['to'])
 
-def nomination_sort(nominations, sort):
 	if sort == 'latest':
-		sorted_instances = NominationInstance.objects.filter(nomination__in=nominations, status=2).order_by('-submitted_at')
+		submissions = submissions.order_by('-submitted_at')
 	elif sort == 'oldest':
-		sorted_instances = NominationInstance.objects.filter(nomination__in=nominations, status=2).order_by('submitted_at')
+		submissions = submissions.order_by('submitted_at')
 
-	return sorted_instances
-	
+	return submissions
 
-def nomination_instance_detail(instance, requested_user):
-	likes_count = instance.likes.count()
-	comments_count = instance.comments.count()
-	instance_detail = {
-		'answers': NominationAnswers.objects.filter(nomination_instance_id=instance, submitted_by_id=instance.user),
-		'submitted_at': str(instance.submitted_at),
-		'is_liked': False,
-		'likes_count': likes_count - 1 if likes_count > 0 else 0,
-		'comments_count': comments_count
+
+def nomination_instance_detail(submission, requested_user):
+	likes_count = submission.likes.count()
+	comments_count = submission.comments.count()
+	is_liked = Like.objects.filter(submission=submission, voter_id=requested_user).exists()
+	instance={
+		'submission': submission,
+		'qa': QuestionAnswers.objects.filter(nomination_submitted=submission),
+		'detail':{
+			'is_liked': is_liked,
+			'likes_count': likes_count - 1 if is_liked else likes_count,
+			'comments_count': comments_count
+		}
 	}
-	for like in instance.likes.all():
-		if like.voter == requested_user:
-			instance_detail['is_liked'] = True
-
-	return instance_detail
-
+	
+	return instance
 
 
 # view for single nomination instance
-def nomination_instance_post(request, nomination_instance_id):
-	instance = NominationInstance.objects.get(id=nomination_instance_id)
-	instance_detail = nomination_instance_detail(instance, request.user)
+def nomination_instance_post(request, nomination_submittion_id):
+	submission = NominationSubmitted.objects.get(id=nomination_submittion_id)
+	instance = nomination_instance_detail(submission, request.user)
 	comment_form = CommentForm()
-	return render(request, 'nominate_app/nomination_instance.html', {'instance': instance, 'detail': instance_detail, 'comment_form':comment_form})
+	return render(request, 'nominate_app/nomination_instance.html', {'instance': instance, 'comment_form':comment_form})
 
 # For Like and unlike in nomination instances
-def nomination_like(request, nomination_instance_id):
-	nomination_instance = get_object_or_404(NominationInstance, id=nomination_instance_id)
-	like, created = Like.objects.get_or_create(voter=request.user, nomination=nomination_instance) 
+def nomination_like(request, nomination_submittion_id):
+	nomination_submission = get_object_or_404(NominationSubmitted, id=nomination_submittion_id)
+	like, created = Like.objects.get_or_create(voter=request.user, submission=nomination_submission) 
 	if created:
 		like.save()
 		return JsonResponse({'value':'like'})
@@ -98,42 +104,62 @@ def nomination_like(request, nomination_instance_id):
 		like.delete()
 		return JsonResponse({'value':'unlike'})
 
+def nomination_liked_people(request, nomination_submittion_id):
+	nomination_submission = get_object_or_404(NominationSubmitted, id=nomination_submittion_id)
+	likes = nomination_submission.likes.all()
+	liked_people = []
+	for like in likes:
+		liked_people.append(like.voter.email)
+
+	return JsonResponse({'liked_people': liked_people})
+
 
 # For comment resource of nomination instance
 class CommentList(View):
-	def get(self, request, nomination_instance_id):
+	def get(self, request, nomination_submittion_id):
 		form = CommentForm()
-		nomination_instance = get_object_or_404(NominationInstance, id=nomination_instance_id)
-		return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'user': request.user, 'form': form, 'instance': nomination_instance}))
+		page = request.GET.get('page', 1)
+		nomination_submission = get_object_or_404(NominationSubmitted, id=nomination_submittion_id)
+		comments = nomination_submission.comments.order_by('-created_date')
+
+		paginator = Paginator(comments, 3)
+
+		try:
+			comments = paginator.page(page)
+		except PageNotAnInteger:
+			comments = paginator.page(1)
+		except EmptyPage:
+			comments = []
+
+		comments.object_list = comments.object_list[::-1]
+
+		return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'user': request.user, 'form': form, 'instance': nomination_submission, 'comments': comments}))
 
 
-	def post(self, request, nomination_instance_id):
-		nomination_instance = get_object_or_404(NominationInstance, id=nomination_instance_id)
+	def post(self, request, nomination_submittion_id):
+		nomination_submission = get_object_or_404(NominationSubmitted, id=nomination_submittion_id)
 		form = CommentForm()
 		form = CommentForm(request.POST)
 		if form.is_valid():
 		  comment = form.save(commit=False)
 		comment = Comment()
-		comment.nomination = nomination_instance
+		comment.submission = nomination_submission
 		comment.text = request.POST['comment']
 		comment.author = request.user
 		comment.save()
-		
-		return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'user': request.user, 'form': form, 'instance': nomination_instance}))
 
-	def delete(self, request, nomination_instance_id, comment_id):
-		nomination_instance = get_object_or_404(NominationInstance, id=nomintion_instance_id)
-		form = CommentForm()
-		comment = get_object_or_404(Comment, id=comment_id)
-		comment.delete()
-		return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'form': form, 'instance': nomination_instance}))
+		comments = []
+		comments.append(comment)
+
+		return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'user': request.user, 'form': form, 'instance': nomination_submission, 'comments': comments}))
 
 
-def comment_remove(request, nomination_instance_id, comment_id):
-	nomination_instance = get_object_or_404(NominationInstance, id=nomination_instance_id)
+def comment_remove(request, nomination_submittion_id, comment_id):
 	form = CommentForm()
 	comment = get_object_or_404(Comment, id=comment_id)
-	comment.delete()
-	return HttpResponse(render_to_string('nominate_app/nomination_comments.html', {'user': request.user, 'form': form, 'instance': nomination_instance}))
+	if comment:
+		comment.delete()
+		return JsonResponse({'status': "deleted"})
 
+	return JsonResponse({'status': "error"})
